requirements.txt
fastapi[all]==0.110.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
google-cloud-firestore==2.20.1
google-genai==1.5.0
